On XMLHttpRequests, add manual timeout to take place of native one as a fallback
We noticed today that on some LG TVs, setting a timeout through a
`XMLHttpRequest`'s `timeout` property didn't have any effect.

The browser version was recent enough, so it seems more of an issue
either due to a bug in the specific browser version integrated there or
linked to an issue brought during the integration of the browser on the
TVs.

Anyway, I propose to add a manual timeout to support the "native" one in
the case where it would fail, to completely avoid issues which can be
problematic enough:

To give an example of such problematic situations, we were investigating
a case where it appeared that some CDN just - relatively rarely - chose
to hang requests forever.
Without any effective timeout solution, this could just lead to infinite
rebuffering as the RxPlayer would just wait for e.g. that segment
request to end before going to the next.

I still chose to rely on the "native" timeout by default and to postpone
the manual one for some seconds, mainly in the perhaps-wrong assumption
that the native one is more precise (also, it might be more efficient,
though here we're speaking of a very rare situation anyway, so this is
not a real argument).