Remove initialVideoBitrate and initialAudioBitrate, add baseBandwidth
This commit removes both `initialAudioBitrate` and `initialVideoBitrate`
APIs, replacing it by just a `baseBandwidth` API.

`baseBandwidth` should be set to a global bandwidth estimate of the
user.

If, for the upcoming v4, the other commits related to bitrate APIs are
all merged, this would have been the last documented APIs forcing the
RxPlayer to separate audio and video bandwidth calculations.

I propose removing it for the following reasons:

  - Applications used it for the wrong reasons for the most part:
    thinking the set value will correlate exactly to the first video or
    audio segments seen by the user.

    In reality, things are much more complex. True bandwidth
    calculations could be performed before pushing the initially-played
    media segments, fast-switching could replace an initially-lower
    quality, factors may be used between that value and the true one
    relied on etc.

  - It finally allows us to remove the separation between audio bandwidth
    and video bandwidth calculations if we want to.

    This may be helpful if we ever support HLS contents, or if we ever
    want to globalize our adaptive logic.

  - Indicating a single bandwidth, instead of one for video and one for
    audio, should be easier to understand as an API.

---

In the RxPlayer code, `baseBandwidth` is then set as the initial
bitrate both for video and audio. As such, the RxPlayer internals do not
change much.

This logic may be problematic however as we should probably take into
account the addition of at least the audio and video Representations
when choosing one, not them individually.
By just using `baseBandwidth` for both individually, we thus may be too
optimistic.

However, this is not as bad as it sounds:

  - There should be factors in play which leads to safer Representation
    choices than just the highest compatible one.

  - For now, applications may consider that fact to set a lower
    `baseBandwidth`.

  - Also, in case of too high bitrates being initially chosen, the
    RxPlayer should realize soon enough that this is the case.

    We may just end-up with the first content taking a little longer
    to play than expected.

Still a better handling of this value should probably be written in the
future.