For stopAtEnd option, rely directly on RxPlayer state changes
We had an issue on LG TV which presumably was linked to a race
condition on this device.

On some LG TVs only, and only when playing two consecutive unencrypted
directfile contents, the second one would never play if:
  - the `stopAtEnd` constructor option was set to `true`
  - the second content was loaded immediately once the RxPlayer switched
    to the `"ENDED"` state.

This was because of a weird race condition, where:
  1. An initial unencrypted directfile content is played until its end.
  2. The `ended` attribute is set to `true` on the media element
  3. Some event (e.g. "timeupdate" or "pause") was then triggered on
     that media element. That's when the RxPlayer first notices that the
     `ended` attribute is set to `true`, switches its state to
     `"ENDED"`, and trigger the corresponding `playerStateChange` event.
  4. On that event, the application loaded the second directfile
     content.
  5. The RxPlayer set on the media element the url of the new
     unencrypted directfile content, as its `src` attribute.
  6. The LG TV's browser decides now, after being back to the event
     loop, to trigger the `"ended"` event for the previous content,
     despite the fact that the content actually changed
  7. The RxPlayer based itself on the `"ended"` event to implement the
     stopAtEnd functionality. Here, it immediately stopped the new
     content, thinking it is the one that just ended (because this is
     the current one).

To work around this issue, I decided to simplify things and just rely
directly on the same internal trigger than for the `"ENDED"` state
change (which is when the `ended` attribute of the media element is
first seen to be set at `true`) to stop the content when `stopAtEnd`
is set to `true`, and not on the `"ended"` event anymore.

The stopping logic happens just after triggering the state change for
the application and only if the state stood as ENDED and if the
corresponding content was not already stopped.