Implement `segmentRequestTimeout` and `manifestRequestTimeout`
I noticed that we still hadn't any way to let an application configures
request timeouts for either the Manifest or the segments.

By default, both of those timeouts are for now set at 30 seconds, which
is a lot. On platforms seen recently where frequent timeouts are occuring
(presumably due to CDN issues), this may easily lead to a rebuffering,
both for live and VoD contents (though sensibly more chance to have one
when playing a live content, as there is less preloaded buffer).

Those high timeouts were set as a safe high value. They could make sense
for large 10+ seconds segments for example.

But on most real encountered cases with 2 to 4 seconds segments, it is
too high and may lead to poor user experience in case of network issues.

---

I though we already allowed its configuration, but turns out we didn't!

Thus I added both a `segmentRequestTimeout` option and a
`manifestRequestTimeout` option to the `networkConfig` optionally given
to a `loadVideo` call.

This also adds the possibility to just remove any timeout by setting the
corresponding option to `-1` (as `undefined` will just lead to the
default 30 seconds timeout used by the RxPlayer).

Though this seems simple, there was still some ambiguous cases:

  - should timeouts be used when an external `manifestLoader` or
    `segmentLoader` is used?

    Here, I chose to just send the timeout as a supplementary property
    to both callbacks. It will be to the application's responsability to
    implement it or not.

    It seemed logic to me as a `manifestLoader` or `segmentLoader` may
    be doing other things than requests or may not want to use timeouts
    in very specific cases.

  - should timeouts be used when loading `"local"` contents?

    I was really not sure here. On one side, I thought that we should
    because the application explicitely set a timeout when loading a
    "local" content as both are given on the same `loadVideo` call.

    On the other side, this would be a change of behavior and most local
    transport are not really usual "requests" per se.

    So for now, I didn't use it there, but I would not be against
    actually using it in the future. Thankfully `"local"` is still
    experimental so we almost can do whatever we want with it ^^

  - What about files that are loaded while parsing a Manifest file, such
    as an DASH UTC Timing's clock or a DASH xlink? Or even more
    ambiguous, a MetaPlaylist subManifest?

    For now, I chose to just use the default 30 seconds timeout here, as I
    wasn't sure at all what to do. Perhaps still another option, but
    naming it would be the hard part here!