Fix redundant endOfStream handling
This commit is a very small fix whose main side-effect should be only to
avoid seeing duplicate endOfStream logs - though it could have been much
worse.

The issue is that in `load_on_media_source.ts`, the observable returned
by the `StreamOrchestrator` was subscribed twice:
  - once by the `ContentTimeBoundariesObserver`, which will listen for
    track changes to re-calculate the current content's duration
  - another time by the whole Observable returned by the
    `loadContentOnMediaSource` function defined in this file, to bubble
    up its event.

Because those two subscriptions are not on a shared Observables,
side-effects in them are provoked twice in a row each time. Depending on
the type of side-effects that could be very problematic. Thankfully, it
seems that the only real side-effects there should be:
  - logs being repeated twice in a row
  - the end-of-stream logic (both to enter and exit) may be executed
    twice in a row, which should not be problematic.
  - the "lockedStream" logic may be repeated twice in a row (fine I
    guess?)
  - discontinuities may be anounced twice in a row (no problem there
    also beside running two times the same logic).

To fix that issue, I chose now to stop relying on the
`StreamOrchestrator` Observable in the `ContentTimeBoundariesObserver`,
relying on our much simpler `SharedReference` concept instead.

In the long term, we plan completely removing RxJS from the code, also
to avoid this kind of subtle bugs that is very easy to introduce.

This bug is a `v3.28.0` regression.