Add autoPlayer `reload` option and improve autoplay default logic in it
Fixes #1159

The `reload` method is a method allowing to load again - and generally
more optimally - the last loaded content, even if the player was stopped
before the `reload` call.

As we know it, this method is mainly used to restart loading a content
even if an unusual and fatal error happened during playback.

By default, this method re-considered all options given to the initial
`loadVideo` call, with the exception of the `startAt` API - where the
last known position was considered instead. It was also possible to
signal a wanted initial position on the `reload` call.

As #1159 suggested, we also should have put more considerations into
the `autoPlay` setting. Whereas until now we only considered the
`autoPlay` setting of the last `loadVideo` call, it seems to make more
sense to respect the last playing state the content was actually last
seen in: if it was paused, we should reload in pause, if it was playing,
we should reload in a playing state.

Rules are actually a little more complicated and depend both on the
player's state and on the current playing condition.

I tried to implement this (in the API part of the code), but the code
I was modifying seemed prone to race conditions: both the playing
conditions and the player's state may rely on each other.

I ended up regrouping the logic updating the player's state, the logic
updating the reloading metadata AND the logic sending position updates
into the same place and in the same order (basically 1. reloading
metadata, 2. state updates and 3. position updates) to put more control
into our hands.

There still remain some "racy" code here and there though:

  - the reloading metadata only bases itself on the current playback
    information, which may not reflect the actual goal (for example if
    a future seek was scheduled but not yet done, we would not know it
    when reloading - we would just reload at the last actual position
    instead).

    This is however relatively minor and can be fixed if it leads to
    actual true issues.

  - If an application seeks then reload synchronously, the reload
    operation would not yet consider that seek.

    Here a solution could be to force `reload` to wait a turn to the
    browser's event loop before actually capturing the last known
    position. However, this could lead to a whole new class of errors.
    I decided here not to handle it for now.

Moreover, a new `autoPlay` property has been added to the `reload`
method, for force an autoPlay behavior.