Fix directfile audiotrack selection on Tizen
We noticed an issue on some Samsung TVs where some (but not all)
directfile contents were playing without sound.

After investigation, it seems that what breaks was the action of
disabling __ALL__ of an `HTMLMediaElement`'s audio tracks before enabling
the one wanted, which is something we did (instead of just disabling the
unwanted tracks first) due to a playback issue on Safari which would
happen without (Safari would just decide to not play).

So basically, an ugly-but-functional-and-still-logically-sound work-around
fixed Safari but broke Tizen.
In the end, I chose to keep the ugly-but-functional-... solution by
default and just perform the original not-ugly solution for Tizen,
because it seems to me there's more chance to re-encounter the Safari
issue (e.g. because another device rely on a similar webkit build) than
re-encounter the Tizen issue (which is especially known to be a broken
target anyway in terms of compat - much more than safari in general) on
a future new device.

Though the more broken target is the only one which profit from the more
elegant logic, which seems counter-intuitive!

I added that in `src/compat` to make it clear that this whole logic is
an ugly compat-oriented work-around.