Avoid staying in LOADING/BUFFERING/SEEKING state in directfile when nothing is buffered
This commit fixes an issue we encountered sometimes with rare directfile
contents, especially on recent Chrome versions and also on the PlayStation
4's WebKit-based one where the RxPlayer would stay in the `LOADING` state
after loading a directfile content, despite the fact that the player is
actively playing the content.

This issue is due to the fact that when playing some directfile
contents, the `buffered` property of the video/audio element stay empty
even when data is obviously buffered as playback is going on.

The RxPlayer then uses that `buffered` property to check if enough data
is buffered to begin playback or go into a `BUFFERING` phase. When
empty, The RxPlayer considers that no data is buffered and that it
should wait until data around the wanted position is before
(re-)starting playback.

After re-checking the WHATWG spec, this browser behavior is not normal
but as we've seen, is actually frequent enough to have been seen in two
separate occasions on two different platforms (chrome desktop and
PlayStation 4).

I chose to provide a work-around for directfile contents by specifically
consider that an empty `buffered` and a `HTMLMediaElement`'s
`readyState` property set to 3 or more, for directfile contents only,
is equivalent to having buffered data for the current time range.
Because a `readyState` superior to 2 (or even equal to 2?) should in
principle not occur when `buffered` has a `0` length, this may have
no bad effect, though as we've seen, the specification does't really
translate what actually can be seen on a browser, so we never really
know.