Re-create MediaKeys at each loadVideo on WebOS
We noticed of an issue on WebOS (LG TV) 2021 and 2022 models where
loading an already-loaded encrypted content would not succeed: it would
load indefinitely.

After investigation, nothing in the JavaScript code seemed to go wrong,
licences were loaded, but nothing was playing.

However, we found out that re-creating the `MediaKeys` instance at each
zap completely fixed the issue, with the cost of potential longer
loading time (not measured yet).

This seems to be definitely an issue linked to LG's software and we will
share to them what we found, but we still chose to preemptively do a
work-around specifically for WebOS 2021 and 2022 models so it works
even know.

If the issue become fixed in the future, we may remove that work-around.

The major part of this commit are added tests on the code handling
content decryption, to ensure that a MediaKeys is only re-created in
some very specific cases (and keep being re-created in those).