Add `updateContentUrls` API
This commit adds the `updateContentUrls` API allowing to update at any
time and during playback, the URL through which a content's Manifest is
reached.

This may be used anytime an application prefer a new Manifest URLs to be
used, whether it is because the previous URL will soon expire, because
the new URL is seen as more qualitative or to rebalance load.

The first argument given to `updateContentUrls` is actually an array of
strings, listing new URLs from the most prioritized to the least
prioritized. This is not used for now, but should soon be used to define
fallback URLs when former one fail to fetch the resource.

`updateContentUrls` can be given a second argument named `refreshNow`
which can be used to ask the RxPlayer to refresh immediately the
Manifest behind the given URL.

For now, this API will throw for directfile contents, I'm not sure
whether it is pertinent to enable it also in this case.