DASH: Now handle `endNumber` DASH attribute
This commit adds handling of the `endNumber` attribute found on some DASH
contents, which is the number the last segment should have.

We decided to add it after seeing it used in the wild, most notably by
some of Canal+ partners.

This attribute is only present in newer 2019+ DASH specifications we
funnily enough did not have access to until now. It was not in the latest
DASH-IF specification neither.