Compat/DRM: generalize the renewal of MediaKeys for WebOS
In the last release, we added a work-around a WebOS (LG TVs) issue [#1175]
where playing multiple times the same encrypted content could lead to
infinite rebuffering.

We thought at the beginning that the issue was only reproducible on
WebOS for 2021 and 2022 TV models. However, we recently reproduced it on a
WebOS 2019 TV, only on different contents - weirdly (the contents triggering
the issue on more recent TV are different from the one triggering it on
older ones).

We're in the process of exchanging with LG on those find but for now, I
chose to just activate the work-around, which may lead to higher loading
time on LG TV, to all kinds of LG TVs.