Add the `getKeySystemConfiguration` method to the RxPlayer
Encryption-related issues are the most frequent category of issues
currently investigated by the RxPlayer team and its associated
applications at least made by Canal+ and our partners.

In that context, we missed a crucial API allowing to facilitate
encryption-related debugging on various platforms:
One that would allow to indicate easily the current key system
configuration relied on.

Particularly the robustness level: PlayReady SL3000 or SL2000? Widevine
L1, L2 or L3? And so on.

The present PR adds the `getKeySystemConfiguration` API which returns
both the **actual** key system string and the
[`MediaKeySystemConfiguration`](https://www.w3.org/TR/encrypted-media/#dom-mediakeysystemconfiguration)
currently relied on.

I put there an emphasis on "actual" because, in opposition to the
`getCurrentKeySystem` API (which I now chose to deprecate, in favor of
the new method), it is [the name actually reported by the
MediaKeySystemAccess](https://www.w3.org/TR/encrypted-media/#dom-mediakeysystemaccess-keysystem)
that is here reported, whereas `getCurrentKeySystem` returned the exact
same string than the `type` originally set on the `keySystems` option -
an arguably less useful value when the RxPlayer could have made translations in
between.