Deprecate methods and options removed in the future v4
To make sure that some API planned to be removed in the incoming v4 aren't
still in use and/or haven't a use we didn't realize until now, I chose to
deprecate some of them in the next v3.30.0 minor version.

The v4 is still planned to stay a long time in beta and we still plan to
maintain the v3 as long as needed for our close partners, but
deprecating options also in the v3 ensure we're not breaking use-cases for
people with not enough resource to make the switch for the v4 beta by
putting more visibility on them (console.warn call, deprecation
documentation etc.).

The following API are here deprecated:
  - The `getVideoLoadedTime` method
  - The `getVideoPlayedTime` method
  - The `transportOptions.aggressiveMode` property of `loadVideo`
  - The `keySystems[].onKeyStatusesChange` callback of `loadVideo`