Add `videoResolutionLimit` constructor option
This new option, which replaces the previous `limitVideoWidth`
constructor option, allows an application to indicate that it wants to
throttle video Representations either to the media element's resolution
or even to the screen's resolution.

This allows to avoid loading bandwidth-heavy video Representations whose
resolution won't be in the end discernible.

What `videoResolutionLimit` brings to the table relative to
`limitVideoWidth` is the possibility to now throttle by the screen's
current resolution - thus handling scenarios where the application want
to stay ready for times where the user decides to enter some kind of
fullscreen mode (whereas `limitVideoWidth` might have taken time going
from the lower resolution when the video was in a small element on the
page to a higher one when set to fullscreen).

This feature has been kind of hastily implemented at the last moment,
due to the realization of a missed opportunity before the official
release of the first v4 beta version. It is unashamedly very inspired
from the shaka-player's `restrictToScreenSize` and
`restrictToElementSize` configs, has I found them to be good ideas!

I hope nonetheless that I didn't forget anything.