Remove RxJS from the demo page
The demo page was the last place where RxJS was remaining.

Removing this dependency from here allows to remove without fear the
dependency as well as various work-arounds that it brought.

The main point here is also to facilitate the readability of our full
demo page, which has been used in the past as an example of an RxPlayer
implementation.

This demo page's codebase is still far from perfect (it would surely
profit from being Redux-based instead of its own state management tool
for readability) but it's a step in the right direction, at least in
my opinion.