DASH: parse `transferCharacteristics` property to better detect hdr
This commit allows to signal HDR AdaptationSets through the API -
thanks to the `hdrInfo` property added in v3.26.0 (june 2021) - when an
`EssentialProperty` or `SupplementalProperty` element with its `schemeIdUri`
attribute is set to `"urn:mpeg:mpegB:cicp:TransferCharacteristics"`.

This is as specified in both the `6.2.5.1` and `10.2.3.4` chapters of
the DASH-IF IOP the way that transfer function differences between
AdaptationSets should be signaled when you have an MPD with both HDR and
SDR Adaptation Sets.

For now, only the "eotf" (the transfer function) is parsed (though we
could also technically parse the `ColourPrimary` and
`MatrixCoefficients` parts of those chapters later).

The goal is to allow an application to detect when multiple
AdaptationSets are present which corresponds to HDR and which
corresponds to SDR video content.