Clean-up cancellation to avoid memory leak
The `v3.30.0` pre-releases and the `v4.0.0-beta.0` beta release has a
memory leak linked to the CancellationSignal abstraction - which is the
new pattern to implement cancellation now that RxJS has been removed.

The idea is that callback to implement clean-up logic can be added to
these `CancellationSignal` which will be triggered in case the
corresponding task is cancelled, exactly like the browser native
`AbortSignal`'s `"abort"` event if you know about it.

However, exactly like a classical event listener that callback is a
closure which may keep unnecessarily in memory variables it contains,
even when the cancellation logic become unneeded, as long as a reference
is somewhere kept on its parent object.

For example, an ended task whose `CancellationSignal` has never been
triggered may still keep alive the callback implementing the cancellation
logic - because that `CancellationSignal` may technically still be
triggered (and thus its logic can still be run) by the task's caller if
it kept a reference to it.

Keeping the cancellation logic "runnable" even after the task has finished
may seem nonsensical, but it has been left that way in some places in the
code because it seemed there were no problem in the corresponding logic in
doing so. Turned out there was, but still not in terms of logic, only in
terms of memory being unnecessarily kept until the `CancellationSignal`
was actually triggered (which generally happens, though might only be in
the long future - for example when the content is stopped).

To fix these, I made sure to always clean-up cancellation logic when the
task is actually finished and the cancellation logic has thus no utility
anymore. I still have some checks to do to see if no memory leak remain.

This is actually an issue we prevented initially when using the
`CancellationSignal`, but I apparently grew more lazy lately! This is
a little sad still as it is an easy footgun.

---

What should be kept in mind and better communicated is that cancellation
logic registered through a `CancellationSignal` has to be treated like
an event listener on an element's whose lifecycle is not known by the
code it is given to (it is linked to its parent `TaskCanceller` much
like an `AbortSignal` is linked to its parent `AbortController`).
Memory may be thus leaked as long as the `TaskCanceller` is alive.