Add createCancellablePromise util
In #1213, I fixed a memory leak by making sure that cancellation logic
was unregistered as soon as possible. Without that work,
cancellation-linked callbacks were most of the time kept in-memory until
the corresponding `CancellationSignal` actually emits, which was
sometimes only after a very-long-lived task (for example when the
corresponding `TaskCanceller` was linked to the playback of the whole
content).

Turned out the most frequent case where we want to unregister
cancellation logic was when using Promises (other cases are very often
long-lived task which don't actually finish until cancellation - those
do not need such clean-up).

When relying on Promises, the cancellation architecture is most of the
time very similar: we generally want to reject if a
`CancellationSignal` emits before the Promise fulfilled, optionally
running some cancellation logic just before doing that, and in cases
where the Promise actually fullfills first, we might want to
unallocate the cancellation logic to free now unnecessary memory.

To facilitate that work and prevent future issues, I added the
`createCancellablePromise` util that does just that, and replaced parts
of the code where it made sense to have it.